Refactor: 불필요한 파일 정리, 제거
