+++ components/Chat/EntryNotice.tsx
