+++ components/Chat/ExitNotice.tsx
