+++ components/Chat/index.ts
@@ -1,8 +1,8 @@
 import ChatLoading from './ChatLoading/ChatLoading';
-import ChatAlert from './ChatAlert1';
+import ChatAlert from './ChatAlert';
 import ChatroomHeader from './ChatroomHeader';
-import EntryNotice from './EntryNotice1';
-import ExitNotice from './ExitNotice1';
+import EntryNotice from './EntryNotice';
+import ExitNotice from './ExitNotice';
 import MyMessage from './MyMessage';
 import OtherMessage from './OtherMessage';
