+++ components/HostList/HostList.tsx
@@ -9,7 +9,7 @@ import styles from '@/components/HostList/HostList1.module.scss';
 import Search from './Search/Search';
 import HostListItem from './HostListItem/HostListItem';
 import HostDetailsModal from './HostDetailsModal/HostDetailsModal';
-import { FirebaseData, Host } from './HostList1.types';
+import { FirebaseData, Host } from './HostList.types';
 
 export default function HostList() {
   const [hostData, setHostData] = useState<Host[]>([]);
