+++ pages/login.tsx
@@ -7,8 +7,8 @@ import { FieldValues, SubmitHandler, useForm } from 'react-hook-form';
 import { useSetRecoilState } from 'recoil';
 import { GetServerSidePropsContext } from 'next';
 import { setToken } from '@/utils/tokenManager';
-import userTokenState from '@/recoil/atoms/userTokenState';
-import { showNavigationState } from '@/recoil/atoms/showNavigationState';
+import userTokenState from '@/stores/atoms/userTokenState';
+import { showNavigationState } from '@/stores/atoms/nav.atoms';
 import styles from '@/components/login.module.scss';
 
 const instance = axios.create({
