+++ stores/atoms/nav.atoms.ts
@@ -1,6 +1,8 @@
 import { atom } from 'recoil';
 
-export const showNavigationState = atom<boolean>({
+const showNavigationState = atom<boolean>({
   key: 'showNavigationState',
   default: true,
 });
+
+export default showNavigationState;
